feat: changed the bottomSheetScaffold for a modalBottomSheet and now it works and it asks you for the dates.
-refactored the huge file Scanner.kt and took some functions into the new files: FoodInputContent.kt and CameraComponents.kt
